requirements.txt
altair==4.2.2
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
blinker==1.5
Bottleneck==1.3.5
branca==0.6.0
brotlipy==0.7.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
DateTime==5.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
flit_core==3.6.0
folium==0.14.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
greenlet==1.1.3
haversine==2.7.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.19.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.16.0
jupyter_client==7.4.8
jupyter-contrib-core==0.4.0
jupyter-contrib-nbextensions==0.7.0
jupyter_core==5.1.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.6.1
jupyter-server==1.23.4
jupyterlab-pygments==0.1.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lxml==4.9.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
msgpack==1.0.4
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.2
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
platformdirs==2.5.2
plotly==5.13.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pynvim==0.4.3
pyOpenSSL==20.0.1
pyparsing==3.0.9
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.2
rich==13.3.1
seaborn==0.12.2
semver==2.13.0
Send2Trash==1.8.0
setuptools==63.2.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.18.1
streamlit-folium==0.11.0
tenacity==8.2.1
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.7.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
watchdog==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.37.1
widgetsnbextension==4.0.5
zipp==3.13.0
zope.interface==5.5.2
